task1/requirements.txt
psycopg2-binary==2.9.10
faker==24.0.0
